added PhoneNumberAlreadyExistsException in teacher-service

+++ Teacher-service/src/main/java/com/brighties/teacherservice/service/TeacherService.java
@@ -3,6 +3,7 @@ package com.brighties.teacherservice.service;
 import com.brighties.teacherservice.dto.TeacherRequestDTO;
 import com.brighties.teacherservice.dto.TeacherResponseDTO;
 import com.brighties.teacherservice.exception.EmailAlreadyExistsException;
+import com.brighties.teacherservice.exception.PhoneNumberAlreadyExistsException;
 import com.brighties.teacherservice.exception.TeacherNotFoundException;
 import com.brighties.teacherservice.mapper.TeacherMapper;
 import com.brighties.teacherservice.model.Teacher;
@@ -37,6 +38,8 @@ public class TeacherService {
 
        if (teacherRepository.existsByEmail(teacherRequestDTO.getEmail())) {
             throw new EmailAlreadyExistsException("Email already exists");
+       }else if (teacherRepository.existsByPhoneNumber(teacherRequestDTO.getPhoneNumber())) {
+           throw new PhoneNumberAlreadyExistsException("Phone number already exists");
        }
         Teacher newTeacher = teacherRepository.save(
                 TeacherMapper.toModel(teacherRequestDTO));
@@ -46,19 +49,30 @@ public class TeacherService {
     }
 
     public TeacherResponseDTO updateTeacher(Long id, TeacherRequestDTO teacherRequestDTO) throws TeacherNotFoundException {
-        Teacher teacher = teacherRepository.findById(id).orElseThrow( () -> new TeacherNotFoundException("teacher not found with id:" + id));
+        Teacher teacher = teacherRepository.findById(id).orElseThrow( () ->
+                new TeacherNotFoundException("teacher not found with id:" + id));
 
         teacher.setName(teacherRequestDTO.getName());
         teacher.setSurname(teacherRequestDTO.getSurname());
         teacher.setAge(teacherRequestDTO.getAge());
+        if (teacherRepository.existsByPhoneNumber(teacherRequestDTO.getPhoneNumber())) {
+            throw new PhoneNumberAlreadyExistsException("Phone number already exists");
+        }
         teacher.setPhoneNumber(teacherRequestDTO.getPhoneNumber());
         teacher.setDescription(teacherRequestDTO.getDescription());
+        if (teacherRepository.existsByEmail(teacherRequestDTO.getEmail())) {
+            throw new EmailAlreadyExistsException("Email already exists");
+        }
         teacher.setEmail(teacherRequestDTO.getEmail());
 
         Teacher updatedTeacher = teacherRepository.save(teacher);
         return TeacherMapper.toDTO(updatedTeacher);
     }
     public void deleteTeacher(Long id) {
+
+        if(!teacherRepository.existsById(id)) {
+            throw new TeacherNotFoundException("Teacher not found with id: " + id);
+        }
         teacherRepository.deleteById(id);
     }
 }
